requirements.txt
streamlit>=1.32.0
tensorflow-cpu>=2.15.0,<2.17.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
